+++ olivia.elmorrison@gmail.com/client/src/App.tsx
@@ -2,8 +2,10 @@
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import { SignupRoute } from "./routes/SignupRoute";
 import { LoginRoute } from "./routes/LoginRoute";
-import { ApplicationRoute } from "./routes/ApplicationRoute";
 import { HomeRoute } from "./routes/HomeRoute";
+import { DashboardRoute } from "./routes/DashboardRoute";
+import { TeamRoute } from "./routes/TeamRoute";
+import { ProjectsRoute } from "./routes/ProjectsRoute";
 
 function App() {
 
@@ -13,8 +15,9 @@ function App() {
           <Route path="/" element={<HomeRoute/>} />
           <Route path="/signup" element={<SignupRoute/>} />
           <Route path="/login" element={<LoginRoute/>} />
-          <Route path="/application" element={<ApplicationRoute/>} />
-          {/* TODO:  */}
+          <Route path="/dashboard" element={<DashboardRoute/>} />
+          <Route path="/team" element={<TeamRoute/>} />
+          <Route path="/projects" element={<ProjectsRoute/>} />
         </Routes>
       </Router>
   );
