--- olivia.elmorrison@gmail.com/client/src/components/Dashboard.tsx
@@ -1,10 +1,0 @@
-import "../styles/Application.css"
-
-export const Dashboard = () => {
-  
-    return (
-      <div className="tab-content">
-        <h1>Dashboard</h1>
-      </div>
-    );
-  }
