--- olivia.elmorrison@gmail.com/client/src/components/Projects.tsx
@@ -1,10 +1,0 @@
-import "../styles/Application.css"
-
-export const Projects = () => {
-  
-    return (
-      <div className="tab-content">
-        <h1>Projects</h1>
-      </div>
-    );
-  }
